Set showing_news_id during updown hotkey, before waiting 100ms.
I was having issues where after using the up/down keys, sometimes
the newsreader panel would show the wrong item, the second-to-last
selected news. Since we wait 100ms before letting the up/down key
populate the newsreader, there's a bit of limbo time which I think
was the problem. This seems to have helped with that.